+++ WebMVC/ex4/microservices/src/main/java/edu/vandy/quoteservices/microservice/hippy/ZippyService.java
@@ -10,7 +10,7 @@ import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
 
@@ -53,7 +53,7 @@ public class ZippyService
         try {
             // Although AtomicInteger is overkill we use it to
             // simplify incrementing the ID in the stream below.
-            AtomicInteger idCount = new AtomicInteger(0);
+            AtomicLong idCount = new AtomicLong(0);
 
             // Convert the filename into a pathname.
             URI uri = ClassLoader
@@ -121,7 +121,7 @@ public class ZippyService
      *                 {@code quoteIds}
      * @return A {@link List} of all requested {@link Quote} objects
      */
-    public List<Quote> getQuotes(List<Integer> quoteIds) {
+    public List<Quote> getQuotes(List<Long> quoteIds) {
         System.out.println("ZippyService.getQuotes()");
 
         if (!mInitialized) {
